*Small changes to README.md *Changed Radius setting to 10 miles

+++ src/GreenOneFoodTrucks.Common/AppSettings.cs
@@ -6,7 +6,7 @@ namespace GreenOneFoodTrucks.Common
     {
         public string AppToken { get; set; }
         public int SoqlQueryLimit { get; set; }
-        public int RadiusOfCentralCoordinateInMeters { get; set; }
+        public double RadiusOfCentralCoordinateInMeters { get; set; }
         public string ResourceId { get; set; }
         public string SanFranciscoFoodTruckApiUrl { get; set; }
     }
